Solved part of issue #1
Boxes moving when beeing dragged towards other box that is blocked,
the bug appears if the blocked box has a higer index than the dragged box.
If the box being drag towards wall with one space between box and player,
a similar bug appears but it does not care about index of box.